moving demo to root level
